+++ src/schemes/TinyBlocks.hpp
@@ -50,8 +50,8 @@ public:
       header_ptr += TinyBlocksSlot::size();
     }
 
-    u32 header_size = header_ptr - dest;
-    u32 payload_size = data_offset - header_size;
+    u64 header_size = header_ptr - dest;
+    u64 payload_size = data_offset - header_size;
     return {header_size, payload_size};
   }
   //---------------------------------------------------------------------------
